requirements.txt
python-chess>=1.999
pygame>=2.5.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
websockets>=12.0
pydantic>=2.0
pytest>=8.0
ruff>=0.4.0
mypy>=1.10.0
